fix(contracts): new stricter typecast rules (#38)
New rules were enforced in
https://github.com/SeismicSystems/seismic-solidity/pull/204, such as not
being allowed to abi.encode a shielded type.

Fixed these mindlessly, but I think we should take some time to look at
the devex related to our current rules. It looks like we force a lot of
manual typecasts which might confuse more than help. For example, we
disallow abi encoding shielded types; I think a more intuitive rule
would be to allow abi encoding shielded types but forcing the output to
be sbytes instead of bytes?

+++ contracts/test/Intelligence.t.sol
@@ -52,8 +52,8 @@ contract IntelligenceTest is Test {
         assertEq(extractCT(encryptedData[1]), directCiphertext[1]);
 
         bytes32[] memory directHashes = new bytes32[](2);
-        directHashes[0] = keccak256(abi.encodePacked(aliceKey));
-        directHashes[1] = keccak256(abi.encodePacked(bobKey));
+        directHashes[0] = keccak256(abi.encodePacked(uint256(aliceKey)));
+        directHashes[1] = keccak256(abi.encodePacked(uint256(bobKey)));
         assertEq(hashes[0], directHashes[0]);
         assertEq(hashes[1], directHashes[1]);
     }
@@ -73,7 +73,7 @@ contract IntelligenceTest is Test {
 
         assertEq(intelligence.numProviders(), 3);
         assertEq(extractCT(encryptedData[2]), directCiphertext);
-        assertEq(hashes[2], keccak256(abi.encodePacked(charlieKey)));
+        assertEq(hashes[2], keccak256(abi.encodePacked(uint256(charlieKey))));
 
         vm.prank(intelligence.owner());
         vm.expectRevert("DUPLICATE_PROVIDER");
@@ -93,7 +93,7 @@ contract IntelligenceTest is Test {
         assertEq(extractCT(encryptedData[0]), directCiphertext);
 
         bytes32 h = directory.keyHash(bob);
-        assertEq(h, keccak256(abi.encodePacked(bobKey)));
+        assertEq(h, keccak256(abi.encodePacked(uint256(bobKey))));
     }
 
     function test_RevertIfRemoveUnkownProvider() public {
